pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gpapi"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for demonstrating payment transactions"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "requests"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]